Split and move into src folder
